Add a confirmation to put a courses into the cart

+++ js/shoppping.js
@@ -17,7 +17,7 @@ let text = '';
 let content = document.getElementById("contentCart");
 for (let i=1; i<= nbObject; i++){
     idCourse = localStorage.getItem(`${i}`)
-    text = text +' '+ COURSES[idCourse].title + '\n';
+    text = text +' '+ COURSES[idCourse].title + '<br>';
 }
 content.innerHTML = text;
 
@@ -33,12 +33,17 @@ function addcart (idCourse){
     }
     //Then, add it
     if (deny == false){
-        nbObject++;
-        addToCart(nbObject,idCourse-1);
-        localStorage.setItem("nbObject", nbObject); //Upgrade the value in the storage
-        location.reload() // Used to refresh the page, else the cart will not apply the current product that we add
+        if ( confirm( "Are you sure to add this ?") ) {
+            nbObject++;
+            addToCart(nbObject,idCourse-1); //Save the ID of the course that we buy
+            localStorage.setItem("nbObject", nbObject); //Upgrade the value in the storage
+            location.reload() // Used to refresh the page, else the cart will not apply the current product that we add
+        }else{
+            console.log("hello");
+        }
     }
 }
+
 function addToCart (nbObject,course){  
     localStorage.setItem(nbObject,course);
     console.log("addcart");
